Update server.py
After forwarding a message, the server sends a delivery confirmation to the sender.
If the recipient doesn't exist, an error message is sent to the sender.     

A broadcast_client_list() function was introduced to send the current list of active clients to all connected clients.
This function is called:
When a new client connects.
When a client disconnects.